Remove some unnecessary casts.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@594018 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/dom/svg/AbstractSVGNormPathSegList.java
@@ -281,7 +281,7 @@ public abstract class AbstractSVGNormPathSegList extends AbstractSVGPathSegList 
 
                         //         Ensure radii are valid
                         if (rx == 0 || ry == 0) {
-                                linetoAbs((float) x, (float) y);
+                                linetoAbs(x, y);
                                 return;
                         }
